modify notification files (NotificationManager, CreatPost, and PostDetail)
